my_tkinter_ai_app/requirements.txt
tensorflow==2.9.1
Pillow==9.2.0
numpy==1.23.0

